+++ Changelog.md
@@ -30,6 +30,7 @@ OpenCore Changelog
 - Fixed ACPI reset register detection in DxeIpl
 - Added MacBookPro16,4 model code
 - Updated builtin firmware versions for SMBIOS and the rest
+- Fixed OSXSAVE reporting when emulating CPUID on newer CPUs
 
 #### v0.5.9
 - Added full HiDPI support in OpenCanopy